+++ app/controllers/recording_studio_presskits/public_press_kits_controller.rb
@@ -2,6 +2,8 @@
 
 module RecordingStudioPresskits
   class PublicPressKitsController < ActionController::Base
+    include RecordingStudio::UsesDefaultLayout
+
     def show
       @press_kit_recording = @parent_recording
       @press_kit = @parent_recordable