+++ app/models/recording_studio_presskits/press_kit.rb
@@ -16,7 +16,8 @@ module RecordingStudioPresskits
     )
     include RecordingStudio::Capabilities::Publishable.to(
       public_controller: "recording_studio_presskits/public_press_kits",
-      public_action: :show
+      public_action: :show,
+      public_layout: "recording_studio/default_layout"
     )
 
     validates :title, presence: true